fix review: better FtConnectImmersionAdvisorDto excluding INDEMNISATION, removed unused bind to convention events
