+++ back/src/domains/convention/use-cases/notifications/NotifyAgencyThatAssessmentIsCreatedWithStatusDidNotShow.unit.test.ts
@@ -6,6 +6,7 @@ import {
   errors,
   expectPromiseToFailWithError,
   type FtConnectIdentity,
+  type FtConnectImmersionAdvisorDto,
   getFormattedFirstnameAndLastname,
   reasonableSchedule,
 } from "shared";
@@ -14,7 +15,6 @@ import {
   type ExpectSavedNotificationsAndEvents,
   makeExpectSavedNotificationsAndEvents,
 } from "../../../../utils/makeExpectSavedNotificationAndEvent.helpers";
-import type { FtConnectImmersionAdvisorDto } from "../../../core/authentication/ft-connect/dto/FtConnectAdvisor.dto";
 import { makeSaveNotificationAndRelatedEvent } from "../../../core/notifications/helpers/Notification";
 import { CustomTimeGateway } from "../../../core/time-gateway/adapters/CustomTimeGateway";
 import {
