+++ back/src/domains/core/authentication/ft-connect/dto/FtConnectAdvisor.dto.ts
@@ -1,27 +1,13 @@
-import type { NotEmptyArray } from "shared";
-
-const ftAdvisorImmersionKinds = ["PLACEMENT", "CAPEMPLOI"] as const;
-export const ftAdvisorKinds = [
-  ...ftAdvisorImmersionKinds,
-  "INDEMNISATION",
-] as const;
+import {
+  type FtConnectImmersionAdvisorsKind,
+  type ftAdvisorKinds,
+  immersionFranceTravailAdvisors,
+} from "shared";
 
 export const isFtAdvisorImmersionKind = (
   input: string,
 ): input is FtConnectImmersionAdvisorsKind =>
   immersionFranceTravailAdvisors.some((value) => value === input);
-
-export const immersionFranceTravailAdvisors: NotEmptyArray<FtConnectImmersionAdvisorsKind> =
-  ["PLACEMENT", "CAPEMPLOI"];
-
-type FtConnectImmersionAdvisorsKind = (typeof ftAdvisorImmersionKinds)[number];
-
-export type FtConnectImmersionAdvisorDto = {
-  email: string;
-  firstName: string;
-  lastName: string;
-  type: FtConnectImmersionAdvisorsKind;
-};
 
 export type FtConnectAdvisorsKind = (typeof ftAdvisorKinds)[number];
